% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleprint.R
\name{print.data.frame}
\alias{print.data.frame}
\alias{format.data.frame}
\title{tibbleprint methods}
\usage{
\method{print}{data.frame}(
  x,
  ...,
  n = NULL,
  width = NULL,
  n_extra = NULL,
  row.names = TRUE,
  base = getOption("tibbleprint.base")
)

\method{format}{data.frame}(x, ..., n = NULL, width = NULL, n_extra = NULL)
}
\arguments{
\item{x}{Object to format or print.}

\item{...}{arguments to be passed to to \code{base::data.frame()}}

\item{n}{Number of rows to show. If \code{NULL}, the default, will print all rows
if less than the \code{print_max} \link[pillar:pillar_options]{option}.
Otherwise, will print as many rows as specified by the
\code{print_min} \link[pillar:pillar_options]{option}.}

\item{width}{Width of text output to generate. This defaults to \code{NULL}, which
means use the \code{width} \link[pillar:pillar_options]{option}.}

\item{row.names}{whether to display row names}

\item{base}{whether to trigger original base behavior}
}
\description{
The arguments are the same as in \code{tibble:::print.tbl} except for \code{row.names},
borrowed from \code{base::print.data.frame}, and \code{base}
which triggers the original behavior. Argument passed to \code{...}
whose names match \code{base::print.data.frame}'s other argument names
("digits", "quote", "right", or "max"), also trigger the
original behavior.
}
